move files from screens to components
