+++ components/ProcessingView.tsx
