+++ components/TranscriptionView.tsx
@@ -1,16 +1,31 @@
+import { StackScreenProps } from "@react-navigation/stack";
 import React from "react";
-import { View, Text } from "react-native";
+import { StyleSheet, View, Text, Button } from "react-native";
+import { RootStackProps } from "../App";
 
-interface TranscriptionViewProps {
-	transcribedText: string;
-}
+type TranscriptionProps = StackScreenProps<RootStackProps, "Transcription">;
 
-const TranscriptionView = (props: TranscriptionViewProps): JSX.Element => {
+const TranscriptionView = ({ navigation, route }: TranscriptionProps): JSX.Element => {
 	return (
-		<View>
-			<Text>{props.transcribedText}</Text>
+		<View style={styles.container}>
+			<Text>Status: {route.params.results}</Text>
+			<Button
+				onPress={() => {
+					navigation.pop(2);
+				}}
+				title="Go Back"
+			/>
 		</View>
 	);
 };
+
+const styles = StyleSheet.create({
+	container: {
+		padding: 10,
+	},
+	filePath: {
+		fontWeight: "bold",
+	},
+});
 
 export default TranscriptionView;
